refactor(web-design): fold fleet namespace into fleet-pi
