+++ web/app/src/lib/pi/session-sidebar-model.test.ts
@@ -1,7 +1,7 @@
 import {
 	sortProjectsByActivity,
 	visibleProjectSessions,
-} from "@prime-agent/web-design/components/fleet/session-sidebar-model";
+} from "@prime-agent/web-design/components/fleet-pi/session-sidebar-model";
 import type { ChatSessionInfo, ProjectSummary } from "@prime-agent/web-protocol";
 import { describe, expect, it } from "vitest";
